backend/requirements.txt
Django==2.2.19
django-colorfield==0.9.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.5.0
Pillow==9.5.0 
psycopg2-binary==2.8.6
python-dotenv==1.0.0
pytest-pythonpath==0.7.3
pytz==2021.1
sqlparse==0.4.1
